Re-aligned package names and files. Updated main version of the application.
