+++ WorldWindowApplicationSample/res/layout/toc_view_dialog.xml
@@ -2,21 +2,19 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent" >
-
-    <nicastel.android.ww.view.DragListView
+    <it.trilogis.android.ww.view.DragListView
         xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tlv="http://schemas.android.com/apk/res/nicastel.android.ww"
+        xmlns:tlv="http://schemas.android.com/apk/res/it.trilogis.android.ww"
         android:id="@android:id/list"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_alignParentTop="false"
         android:drawSelectorOnTop="false"
+        tlv:dragndrop_background="#99ffbb33"
         tlv:grabber="@+id/toc_item_grabber"
         tlv:grabber_alignment="right"
         tlv:normal_height="46dip"
-        tlv:dragndrop_background="#99ffbb33"
         tlv:remove_mode="trash" />
-
     <ImageView
         android:id="@+id/removeView"
         android:layout_width="match_parent"
@@ -27,5 +25,4 @@
         android:src="@android:drawable/ic_delete"
         android:visibility="invisible" >
     </ImageView>
-
 </RelativeLayout>
