+++ WorldWindowApplicationSample/src/it/trilogis/android/ww/dialogs/TocDialog.java
@@ -13,17 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nicastel.android.ww.dialogs;
+package it.trilogis.android.ww.dialogs;
 
 import gov.nasa.worldwind.WorldWindowGLSurfaceView;
 import gov.nasa.worldwind.layers.Layer;
 import gov.nasa.worldwind.layers.LayerList;
-import nicastel.android.ww.R;
-
+import it.trilogis.android.ww.R;
+import it.trilogis.android.ww.dialogs.AddWMSDialog.OnAddWMSLayersListener;
+import it.trilogis.android.ww.view.DragListView;
 import java.util.List;
-
-import nicastel.android.ww.dialogs.AddWMSDialog.OnAddWMSLayersListener;
-import nicastel.android.ww.view.DragListView;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.Dialog;
